refactor: add folders for modules
