+++ rss-puzzle/src/components/main/checkResult/checkResultOrder.ts
