+++ rss-puzzle/src/components/main/checkResult/checkSentence.ts
