+++ rss-puzzle/src/components/main/definitions/interfaces.ts
