+++ rss-puzzle/src/components/start/start.css
@@ -80,7 +80,7 @@
 }
 
 body {
-  background: url(bg.jpg) no-repeat center center fixed;
+  background: url(./images/bg.jpg) no-repeat center center fixed;
   -webkit-background-size: cover;
   -moz-background-size: cover;
   -o-background-size: cover;
